LayoutTests/media/media-source/media-source-append-media-segment-without-init-expected.txt


RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(samples))
EVENT(error)
END OF TEST
